requirements.txt
pandas==2.2.0
openpyxl==3.11.0
fpdf2==2.8.1
streamlit==1.36.0
